Add first test for seconds row.
We're going slightly off-spec this time to implement something that might be more useful. Instead of being a converter for Berlin/Digital time, we're going to create a BerlinClock that reports the actual Berlin time at this current instant, which will save the client the hassle of getting the current time in the correct format and passing it in for conversion. We'll see where it takes us, anyway.